+++ packages/pangea-node-sdk/src/services/index.ts
@@ -2,7 +2,7 @@ import AuditService from "./audit.js";
 import EmbargoService from "./embargo.js";
 import BaseService from "./base.js";
 import RedactService from "./redact.js";
-import { FileIntelService, DomainIntelService } from "./intel.js";
+import { FileIntelService, DomainIntelService, IPIntelService, URLIntelService } from "./intel.js";
 
 export default {
   AuditService,
@@ -11,4 +11,6 @@ export default {
   RedactService,
   FileIntelService,
   DomainIntelService,
+  IPIntelService,
+  URLIntelService,
 };