+++ packages/pangea-node-sdk/tests/integration/intel.test.ts
@@ -1,12 +1,15 @@
 import PangeaConfig from "../../src/config";
-import { FileIntelService, DomainIntelService } from "../../src/services/intel";
 import { it, expect } from "@jest/globals";
+import { TestEnvironment, getTestDomain, getTestToken } from "../../src/utils/utils";
+import { FileIntelService, DomainIntelService, IPIntelService, URLIntelService } from "../../src";
 
-const token = process.env.PANGEA_INTEGRATION_TOKEN || "";
-const testHost = process.env.PANGEA_INTEGRATION_DOMAIN || "";
+const token = getTestToken(TestEnvironment.LIVE);
+const testHost = getTestDomain(TestEnvironment.LIVE);
 const config = new PangeaConfig({ domain: testHost });
 const fileIntel = new FileIntelService(token, config);
 const domainIntel = new DomainIntelService(token, config);
+const ipIntel = new IPIntelService(token, config);
+const urlIntel = new URLIntelService(token, config);
 
 it("file lookup should succeed", async () => {
   const options = { provider: "reversinglabs", verbose: true, raw: true };
@@ -21,9 +24,46 @@ it("file lookup should succeed", async () => {
   expect(response.result.data.verdict).toBe("malicious");
 });
 
+it("file lookup with filepath should succeed", async () => {
+  const options = { provider: "reversinglabs", verbose: true, raw: true };
+  const response = await fileIntel.lookupFilepath("./README.md", options);
+  expect(response.status).toBe("Success");
+  expect(response.result.data).toBeDefined();
+  expect(response.result.data.verdict).toBe("benign");
+});
+
+it("file lookup with filepath should faild", async () => {
+  const options = { provider: "reversinglabs", verbose: true, raw: true };
+
+  try {
+    const response = await fileIntel.lookupFilepath("./not/a/real/path/file.txt", options);
+  } catch (e: unknown) {
+    // @ts-ignore
+    expect(e.code).toBe("ENOENT");
+  }
+});
+
 it("Domain lookup should succeed", async () => {
   const options = { provider: "domaintools", verbose: true, raw: true };
   const response = await domainIntel.lookup("737updatesboeing.com", options);
+
+  expect(response.status).toBe("Success");
+  expect(response.result.data).toBeDefined();
+  expect(response.result.data.verdict).toBe("malicious");
+});
+
+it("IP lookup should succeed", async () => {
+  const options = { provider: "crowdstrike", verbose: true, raw: true };
+  const response = await ipIntel.lookup("93.231.182.110", options);
+
+  expect(response.status).toBe("Success");
+  expect(response.result.data).toBeDefined();
+  expect(response.result.data.verdict).toBe("malicious");
+});
+
+it("URL lookup should succeed", async () => {
+  const options = { provider: "crowdstrike", verbose: true, raw: true };
+  const response = await urlIntel.lookup("http://113.235.101.11:54384", options);
 
   expect(response.status).toBe("Success");
   expect(response.result.data).toBeDefined();