+++ packages/react-shared/src/components/PangeaDataGrid/index.tsx
@@ -25,7 +25,10 @@ import SearchBar from "./components/Search";
 
 import { ConditionalOption } from "../ConditionalAutocomplete";
 import { PreviewPanel } from "./components/PreviewPanel";
-import { FilterFormProps } from "./components/Search/FiltersForm";
+import {
+  FilterFormProps,
+  FilterOptions,
+} from "./components/Search/FiltersForm";
 import { Visibility } from "./components/Search/ColumnsPopout";
 
 export interface PangeaDataGridProps<
@@ -42,6 +45,7 @@ export interface PangeaDataGridProps<
   ServerPagination?: {
     page: number;
     pageSize: number;
+    paginationRowCount?: number;
     rowCount: number;
     onPageChange: (page: number) => void;
     onPageSizeChange?: (pageSize: number) => void;
@@ -52,6 +56,7 @@ export interface PangeaDataGridProps<
   };
   Search?: {
     query?: string;
+    placeholder?: string;
     onChange: (query: string) => void;
     conditionalOptions?: ConditionalOption[];
     Filters?: FilterFormProps<FiltersObj>;
@@ -309,5 +314,7 @@ const PangeaDataGrid = <
     </div>
   );
 };
+
+export type { FilterOptions, FilterFormProps };
 
 export default PangeaDataGrid;