+++ packages/pangea-node-sdk/CHANGELOG.md
@@ -5,11 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.1.1] - 2023-01-05
+
+### Fixed
+
+- Exported IP and URL intel service clients
+
+### Changed
+
+- Updated doc strings for intel services
+- Made the `options` parameter in intel services an optional parameter
+
+## [1.1.0] - 2023-01-05
 
 ### Added
 
 - This CHANGELOG
+- Functions to get token and domain according to test environment (PROD/DEV/STG)
 
 ### Fixed
 
@@ -34,5 +46,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Domain Intel client
 - Redact client
 
-[unreleased]: https://github.com/pangeacyber/pangea-javascript/compare/v1.0.0...main
+[unreleased]: https://github.com/pangeacyber/pangea-javascript/compare/v1.1.0...main
+[1.1.0]: https://github.com/pangeacyber/pangea-javascript/compare/v1.0.0...v1.1.0
 [1.0.0]: https://github.com/pangeacyber/pangea-javascript/releases/tag/v1.0.0